refactor, use sdk create_queue to get better performance
